Process __package.rb files on the fast path
We already guard __package.rb file changes from the fast path, but if
the file hasn't changed allow it to go through. However, we would also
filter __package.rb files out of the updated file set before running the
pipeline, which would lead to running the fast path on zero files.
Instead, let's build up the package files in a separate vector that we
extend with related package files, and merge right before the pipeline.